@@ -3,20 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "5.16.1",
-    "@testing-library/react": "12.1.2",
-    "@testing-library/user-event": "13.5.0",
-    "axios": "0.25.0",
-    "date-fns": "2.28.0",
-    "final-form": "4.20.6",
-    "packages": "0.0.8",
-    "react": "17.0.2",
-    "react-big-calendar": "0.38.5",
-    "react-dom": "17.0.2",
-    "react-final-form": "6.5.7",
-    "react-router-dom": "6.2.1",
-    "react-scripts": "5.0.0",
-    "web-vitals": "2.1.3"
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.3.3",
+    "date-fns": "^2.29.3",
+    "final-form": "^4.20.9",
+    "packages": "^0.0.8",
+    "react": "^18.2.0",
+    "react-big-calendar": "^1.6.8",
+    "react-dom": "^18.2.0",
+    "react-final-form": "^6.5.9",
+    "react-router-dom": "^6.8.1",
+    "react-scripts": "^5.0.1",
+    "web-vitals": "^3.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -43,9 +43,9 @@
     ]
   },
   "devDependencies": {
-    "autoprefixer": "10.4.2",
+    "autoprefixer": "^10.4.13",
     "cypress-file-upload": "5.0.8",
-    "postcss": "8.4.5",
-    "tailwindcss": "3.0.15"
+    "postcss": "^8.4.21",
+    "tailwindcss": "^3.2.7"
   }
 }
